requirements.txt
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.70
botocore==1.42.70
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
deprecated==1.3.1
distro==1.9.0
et-xmlfile==2.0.0
google-auth==2.49.1
google-genai==1.67.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
img2pdf==0.6.3
iniconfig==2.3.0
jmespath==1.1.0
lxml==6.0.2
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pikepdf==10.5.0
pillow==12.1.1
pluggy==1.6.0
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.2
pymupdf==1.27.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
websockets==16.0
wrapt==2.1.2
